+++ Astral One Engine/Engine/Core/Influence Maps/InfluenceMapCalculator.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 public protocol InfluenceMapCalculator {
-    func getInfluenceMap() -> [[Float]]
+    func getInfluenceMap() -> [[Double]]
 }
